session-management/pyproject.toml
[project]
name = "function-calling-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.9",
    "python-decouple>=3.8",
]
